Native socket handler construction fix

+++ tests/Spiffe.Tests/Integration/TestIntegration.Unix.cs
@@ -1,9 +1,6 @@
-#if !OS_WINDOWS
-
 using System.ComponentModel;
-using FluentAssertions;
-using Grpc.Net.Client;
-using Spiffe.Grpc;
+using System.Runtime.InteropServices;
+using Xunit;
 
 namespace Spiffe.Tests.Integration;
 
@@ -16,9 +13,13 @@ public partial class TestIntegration
     [Category(Constants.Integration)]
     public async Task TestFetchViaUnixSocket()
     {
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+        {
+            return;
+        }
+
         string socket = Path.Join(Path.GetTempPath(), $"workload-api-{Guid.NewGuid()}.sock");
         await RunTest($"unix://{socket}");
         File.Delete(socket);
     }
 }
-#endif
